refactor more, take 2
